+++ src/components/ol/util.ts
@@ -6,6 +6,7 @@
 
 import { default as OlMap } from "ol/Map";
 import { default as OlLayer } from "ol/layer/Layer";
+import OlTileLayer from "ol/layer/Tile";
 
 export function findMapLayer(map: OlMap, layerId: string): OlLayer | null {
   const layerGroup = map.getLayers();
@@ -16,4 +17,14 @@ export function findMapLayer(map: OlMap, layerId: string): OlLayer | null {
     }
   }
   return null;
+}
+
+export function findDatasetMapLayer(map: OlMap): OlLayer | null {
+  // Find the first active dataset tile layer named "variable", "rgb",
+  // "variable2", "rgb2" (in that order)
+  const layer = (["variable", "rgb", "variable2", "rgb2"] as const)
+    .map((name) => findMapLayer(map, name))
+    .find((layer) => layer instanceof OlTileLayer);
+
+  return layer as OlLayer;
 }